client/requirements.txt
pyautogui==0.9.54
requests==2.31.0
pillow==10.0.1
python-dotenv==1.0.0
pytesseract==0.3.10
opencv-python==4.8.1.78
imagehash==4.3.1
PyGetWindow==0.0.9
keyboard==0.13.5
pystray==0.19.5
